Step&Choice Controller + Form (rajout des exceptions pour l'upload de fichier)

+++ src/php/Controller/ChoiceController.php
@@ -9,39 +9,64 @@ use \Server\Response;
 
 class ChoiceController {
   public static function addChoice() {
-    $answer = Form::getField("answer");
-    $idStep = Form::getField("idStep");
-    $transitionText = Form::getField("transitionText");
-    $idNextStep = Form::getField("idNextStep");
+    $answer = Form::getField("Answer");
+    $idStep = Form::getField("IDStep");
+    $transitionText = Form::getField("TransitionText");
+    $idNextStep = Form::getField("IDNextStep");
     if ($answer == NULL || $idStep == NULL || $transitionText == NULL || $idNextStep == NULL)
-    return Response::jsonResponse(array(
-      'status' => "error",
-      'message' => "ta race tu ne peux pas ajouter ce choix !"
-    ));
+      return Response::jsonResponse(array(
+        'status' => "error",
+        'message' => "ta race tu ne peux pas ajouter ce choix !"
+      ));
     $choice = new Choice($answer, $idStep, $transitionText, $idNextStep);
     $choice->save();
     Response::jsonResponse($choice);
   }
 
   public static function updateChoice() {
-    $data = Form::getForm();
+    $data = Form::getFullForm(); //si l'id n'est pas présent, on retourne null
+    if(!isset($data["IDChoice"]) || $data["IDChoice"]== null ){
+      Response::jsonResponse(array(
+        'status' => "error",
+        'message' => "Choix invalide"
+      ));
+      return null;
+    }
+
     $entries = array();
-    $fields = array_keys($data);
+    $fields = array("IDChoice","Answer","IDStep","TransitionText","IDNextStep");
     foreach ($fields as $field) {
-      if($data[$field] == NULL) {
-        $entries[$field]=" ";
+      if(!isset($data[$field]) || $data[$field]=="") {
+        $entries[$field]="";
+      }
+      else {
+        $entries[$field]=$data[$field];
       }
     }
-    $choice = new Choice($answer, $idStep, $transitionText, $idNextStep);
-    $choice->id = "id";
+    $choice = new Choice($entries["Answer"], $entries["IDStep"], $entries["TransitionText"],$entries["IDNextStep"]);
+    $choice->id = $entries["IDChoice"];
+
     $choice->update($entries);
+    return 1;
   }
 
   public static function deleteChoice() {
-    $id = Form::getField("id");
-    $choice = new Choice("", "", "", "");
-    $choice->id = $id;
-    $choice->delete();
+    $id = Form::getField("IDChoice");
+    if(!$id){
+      return Response::jsonResponse(array(
+        'status' => "error",
+        'message' => "Suppression de choix invalide"
+      ));
+    }
+    else{
+      $choice = new Choice("", 0, "", "",0);
+      $choice->id = $id;
+      $choice->delete();
+      return Response::jsonResponse(array(
+        'status' => "ok",
+        'message' => "Le choix a été supprimé"
+      ));
+    }
   }
 
 }
